Fix small shifts (use double)

+++ native/src/common2.h
@@ -76,8 +76,8 @@ private :
 	float density;
 	float screenDensityRatio;
 
-	float leftX;
-	float topY;
+	double leftX;
+	double topY;
 	int width;
 	int height;
 	int defaultColor;
@@ -109,7 +109,7 @@ public:
 public :
 
 	// calculated
-	float tileDivisor;
+	double tileDivisor;
 	float cosRotateTileSize;
 	float sinRotateTileSize;
 
@@ -153,7 +153,7 @@ public:
 		this->tileDivisor = (1 << (31 - z));
 	}
 
-	void setTileDivisor(float tileDivisor) {
+	void setTileDivisor(double tileDivisor) {
 		this->tileDivisor = tileDivisor;
 	}
 
@@ -205,11 +205,11 @@ public:
 		return zoom;
 	}
 
-	inline float getLeft() {
+	inline double getLeft() {
 		return leftX;
 	}
 
-	inline float getTop() {
+	inline double getTop() {
 		return topY;
 	}
